test(control-ui): split terminal readiness coverage
