+++ ui/src/components/terminal/terminal-panel.test.ts
@@ -104,7 +104,7 @@ async function startPanelWithPendingOpen() {
   await vi.waitFor(() =>
     expect(requests.some(({ method }) => method === "terminal.open")).toBe(true),
   );
-  return { panel, createOptions: createOptions!, open, requests };
+  return { createOptions: createOptions!, open, requests };
 }
 
 describe("OpenClawTerminalPanel", () => {
@@ -223,22 +223,6 @@ describe("OpenClawTerminalPanel", () => {
     ]);
   });
 
-  it("shows a connecting animation while a terminal open is in flight", async () => {
-    const { panel, open } = await startPanelWithPendingOpen();
-    await panel.updateComplete;
-
-    expect(panel.renderRoot.querySelector(".tp-connecting")?.textContent).toContain(
-      "Connecting to session",
-    );
-    expect(panel.renderRoot.querySelector(".tp-tab")?.classList.contains("is-connecting")).toBe(
-      true,
-    );
-
-    open.resolve(terminalOpenResult("session-1"));
-    await vi.waitFor(() => expect(panel.renderRoot.querySelector(".tp-connecting")).toBeNull());
-    expect(panel.renderRoot.querySelector(".tp-tab")?.classList.contains("is-live")).toBe(true);
-  });
-
   it("drops whole startup input chunks beyond the pending-input cap", async () => {
     const { createOptions, open, requests } = await startPanelWithPendingOpen();
     const accepted = "a".repeat(8 * 1024);
@@ -269,151 +253,6 @@ describe("OpenClawTerminalPanel", () => {
       );
     });
     expect(requests.some(({ method }) => method === "terminal.input")).toBe(false);
-  });
-
-  it("opens a new titled tab for a catalog toggle request", async () => {
-    const controller = createTerminalController();
-    createGhosttyTerminalMock.mockResolvedValue(controller);
-    const requests: Array<{ method: string; params: unknown }> = [];
-    let listener: ((event: { event: string; payload: unknown }) => void) | undefined;
-    const client: TerminalGatewayClient = {
-      forceReconnect: () => {},
-      request: async <T>(method: string, params?: unknown) => {
-        requests.push({ method, params });
-        return {
-          ...terminalOpenResult("catalog-terminal-1"),
-          title: "codex resume 0d5c…",
-        } as T;
-      },
-      addEventListener: (nextListener) => {
-        listener = nextListener;
-        return () => {
-          listener = undefined;
-        };
-      },
-    };
-    const panel = document.createElement(TERMINAL_PANEL_ELEMENT_NAME) as OpenClawTerminalPanel;
-    panel.client = client;
-    panel.available = true;
-    document.body.append(panel);
-    const catalog = { catalogId: "codex", hostId: "node:mac", threadId: "thread" };
-
-    panel.handleToggleRequest(new CustomEvent("openclaw:terminal-toggle", { detail: { catalog } }));
-
-    await vi.waitFor(() => {
-      expect(requests).toContainEqual({
-        method: "terminal.open",
-        params: { agentId: undefined, cols: 100, rows: 30, catalog },
-      });
-    });
-    await panel.updateComplete;
-    expect(panel.renderRoot.querySelector(".tp-tab")?.textContent).toContain("codex resume 0d5c…");
-    expect(panel.renderRoot.querySelector(".tp-connecting")?.textContent).toContain(
-      "Connecting to session",
-    );
-
-    listener?.({
-      event: "terminal.data",
-      payload: { sessionId: "catalog-terminal-1", seq: 5, data: "ready" },
-    });
-    await vi.waitFor(() => expect(panel.renderRoot.querySelector(".tp-connecting")).toBeNull());
-    expect(new TextDecoder().decode(controller.write.mock.calls[0]?.[0])).toBe("ready");
-    expect(sessionStorage.getItem("openclaw.terminal.sessions.v1")).toBe(
-      JSON.stringify(["catalog-terminal-1"]),
-    );
-  });
-
-  it("marks a catalog terminal ready when its first visible output is a replay", async () => {
-    const controller = createTerminalController();
-    createGhosttyTerminalMock.mockResolvedValue(controller);
-    const requests: Array<{ method: string; params: unknown }> = [];
-    let listener: ((event: { event: string; payload: unknown }) => void) | undefined;
-    const client: TerminalGatewayClient = {
-      forceReconnect: () => {},
-      request: async <T>(method: string, params?: unknown) => {
-        requests.push({ method, params });
-        if (method === "terminal.attach") {
-          return {
-            ...terminalOpenResult("catalog-terminal-1"),
-            buffer: "recovered output",
-            seq: 12,
-          } as T;
-        }
-        return {
-          ...terminalOpenResult("catalog-terminal-1"),
-          title: "claude --resume 1234…",
-        } as T;
-      },
-      addEventListener: (nextListener) => {
-        listener = nextListener;
-        return () => {
-          listener = undefined;
-        };
-      },
-    };
-    const panel = document.createElement(TERMINAL_PANEL_ELEMENT_NAME) as OpenClawTerminalPanel;
-    panel.client = client;
-    panel.available = true;
-    document.body.append(panel);
-
-    panel.handleToggleRequest(
-      new CustomEvent("openclaw:terminal-toggle", {
-        detail: { catalog: { catalogId: "anthropic", hostId: "node:mac", threadId: "thread" } },
-      }),
-    );
-    await vi.waitFor(() => expect(panel.renderRoot.querySelector(".tp-connecting")).not.toBeNull());
-
-    listener?.({
-      event: "terminal.data",
-      payload: { sessionId: "catalog-terminal-1", seq: 12, data: "gap" },
-    });
-
-    await vi.waitFor(() => expect(panel.renderRoot.querySelector(".tp-connecting")).toBeNull());
-    expect(requests).toContainEqual({
-      method: "terminal.attach",
-      params: { sessionId: "catalog-terminal-1" },
-    });
-    expect(controller.terminal.reset).toHaveBeenCalledOnce();
-    expect(new TextDecoder().decode(controller.write.mock.calls[0]?.[0])).toBe("recovered output");
-  });
-
-  it("closes a catalog terminal and shows an error when no output arrives", async () => {
-    createGhosttyTerminalMock.mockResolvedValue(createTerminalController());
-    const requests: Array<{ method: string; params: unknown }> = [];
-    const client: TerminalGatewayClient = {
-      forceReconnect: () => {},
-      request: async <T>(method: string, params?: unknown) => {
-        requests.push({ method, params });
-        return (
-          method === "terminal.open"
-            ? { ...terminalOpenResult("catalog-terminal-1"), title: "claude --resume 1234…" }
-            : {}
-        ) as T;
-      },
-      addEventListener: () => () => {},
-    };
-    const panel = document.createElement(TERMINAL_PANEL_ELEMENT_NAME) as OpenClawTerminalPanel;
-    panel.client = client;
-    panel.available = true;
-    (panel as unknown as { catalogReadyTimeoutMs: number }).catalogReadyTimeoutMs = 5;
-    document.body.append(panel);
-
-    panel.handleToggleRequest(
-      new CustomEvent("openclaw:terminal-toggle", {
-        detail: { catalog: { catalogId: "anthropic", hostId: "node:mac", threadId: "thread" } },
-      }),
-    );
-
-    await vi.waitFor(() => {
-      expect(panel.renderRoot.querySelector(".tp-error")?.textContent).toContain(
-        "Session did not connect within 30 seconds",
-      );
-    });
-    expect(requests).toContainEqual({
-      method: "terminal.close",
-      params: { sessionId: "catalog-terminal-1" },
-    });
-    expect(panel.renderRoot.querySelector(".tp-tab")).toBeNull();
   });
 
   it("reattaches persisted sessions before opening a catalog tab", async () => {
